UXD-338 Fix: move variations
